Added character image coding
Added and un-commented the coding for all of the character on-screen
images throughout the scripting up to the end of the day_01 script. I
will work on the rest of the script later, and will eventually have to
add the locations for the character image show statements. Also added a
dummy character image.