Added text speed and click to continue icon
Added a text display speed of 40. Added a click to continue icon that
displays, fading in and out, after a character finishes their dialogue.
Also added a new transient layer called "icons" that is used by the
character heart meters, as the click to contiune function would wipe the
previously used "overlay" layer.